[patch] change IntegerFieldRecord.value type from integer to float(double)
[fix] remove non alphanumeric, whitespace, -, _ character from xls(spredsheet).sheet title

+++ src/Query/Domain/Model/Shared/FormRecord.php
@@ -274,7 +274,7 @@ class FormRecord
         $fieldRecord = $this->stringFieldRecords->filter($p)->first();
         return empty($fieldRecord)? null : $fieldRecord->getValue();
     }
-    public function getIntegerFieldRecordValueCorrespondWith(IntegerField $integerField): ?int
+    public function getIntegerFieldRecordValueCorrespondWith(IntegerField $integerField): ?float
     {
         $p = function (IntegerFieldRecord $integerFieldRecord) use ($integerField) {
             return $integerFieldRecord->isActiveFieldRecordCorrespondWith($integerField);
